Refactor homepage.js, remove DanceClass.js
